refactor(route): 旧ViewModel削除の準備を完了
